requirements.txt
streamlit
ultralytics
fer==22.4.0
tensorflow==2.20.0
tf-keras
opencv-python-headless
pillow
numpy
firebase-admin
pandas
python-multipart
cloudinary
plotly
gradio
sqlalchemy
pymysql
fastapi
uvicorn[standard]